mod 158-b

+++ src/main/kotlin/158/b.kt
@@ -1,6 +1,6 @@
 import java.lang.Math
 
 fun main(args: Array<String>) {
-    val (n,a,b) = readLine()!!.split(" ").map { it.toInt() }
+    val (n,a,b) = readLine()!!.split(" ").map { it.toLong() }
     println(n / (a + b) * a + Math.min(a, n % (a + b)))
 }
